rename recent-items-handler -> udt-helper
