+++ src/components/LocalStorageApp/index.js
@@ -15,7 +15,7 @@ import { setIn } from "seamless-immutable"
 import AppErrorBoundary from "../AppErrorBoundary"
 import useEventCallback from "use-event-callback"
 import usePreventNavigation from "../../utils/use-prevent-navigation"
-import jsonHandler from "../../utils/file-handlers/recent-items-handler"
+import jsonHandler from "../../utils/file-handlers/udt-helper"
 import UpdateAWSStorage from "../../utils/file-handlers/update-aws-storage"
 const randomId = () => Math.random().toString().split(".")[1]
